requirements.txt
azure-core
azure-ai-inference
azure-identity
python-dotenv
flask